requirements.txt
# Django
Django==1.6.7
MySQL-python==1.2.5
redis==2.9.1
django-redis==3.5.1
django-redis-sessions==0.4.0
South==0.8.4
django-extensions==1.3.3
django_compressor==1.3
django-grappelli==2.5.1
django-tastypie==0.11.0
django-devserver==0.8.0

# Python generic
path.py==5.1
iso8601==0.1.8
pytz==2014.4
requests==2.2.1

django-appconf==0.6
python-dateutil==2.2
